requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
filelock==3.8.2
Flask==2.2.2
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.3
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parsel==1.7.0
prompt-toolkit==3.0.36
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
queuelib==1.6.2
redis==4.4.0
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.5
Werkzeug==2.2.2
XlsxWriter==3.0.3
yarl==1.8.2
zope.interface==5.5.2
